+++ seqra-jvm-sast-dataflow/src/main/kotlin/org/seqra/jvm/sast/dataflow/DebugOptions.kt
@@ -6,4 +6,5 @@ data class DebugOptions(
     val factReachabilitySarif: Boolean,
     val enableVulnSummary: Boolean,
     val runRuleTests: Boolean,
+    val debugRunAnalysisOnSelectedEntryPoints: String?
 )