+++ src/main/kotlin/org/seqra/jvm/sast/project/ProjectEntryPointsSelector.kt
@@ -6,20 +6,37 @@ import org.seqra.jvm.sast.project.spring.springWebProjectEntryPoints
 
 private val logger = object : KLogging() {}.logger
 
-fun ProjectAnalysisContext.selectProjectEntryPoints(): List<JIRMethod> = getEntryPoints()
+fun ProjectAnalysisContext.selectProjectEntryPoints(options: ProjectAnalysisOptions): List<JIRMethod> =
+    getEntryPoints(options)
 
-private fun ProjectAnalysisContext.getEntryPoints(): List<JIRMethod> {
+private fun ProjectAnalysisContext.getEntryPoints(options: ProjectAnalysisOptions): List<JIRMethod> {
     logger.info { "Search entry points for project: ${project.sourceRoot}" }
     val springEp = springWebProjectContext?.springWebProjectEntryPoints().orEmpty()
     return when (projectKind) {
-        ProjectKind.UNKNOWN -> allProjectEntryPoints() + springEp
+        ProjectKind.UNKNOWN -> allProjectEntryPoints(options) + springEp
         ProjectKind.SPRING_WEB -> springEp
     }
 }
 
-private fun ProjectAnalysisContext.allProjectEntryPoints(): List<JIRMethod> =
-    projectClasses.projectPublicClasses()
-        .flatMapTo(mutableListOf()) { it.publicAndProtectedMethods() }
-        .also {
-            it.sortWith(compareBy<JIRMethod> { it.enclosingClass.name }.thenBy { it.name })
-        }
+private fun ProjectAnalysisContext.allProjectEntryPoints(options: ProjectAnalysisOptions): List<JIRMethod> {
+    val debugEp = options.debugOptions?.debugRunAnalysisOnSelectedEntryPoints
+    if (debugEp == null) {
+        return projectClasses.projectPublicClasses()
+            .flatMapTo(mutableListOf()) { it.publicAndProtectedMethods() }
+            .ordered()
+    }
+
+    val allMethods = projectClasses.projectAllAnalyzableClasses()
+        .flatMapTo(mutableListOf()) { it.allAnalyzableMethods() }
+        .ordered()
+
+    if (debugEp == "*") return allMethods
+
+    val (clsName, methodName) = debugEp.split('#')
+    return allMethods.filter { it.name == methodName && it.enclosingClass.name == clsName }
+}
+
+private fun MutableList<JIRMethod>.ordered(): List<JIRMethod> {
+    sortWith(compareBy<JIRMethod> { it.enclosingClass.name }.thenBy { it.name })
+    return this
+}